+++ Tests/StreamChatTests/ChatClient_Tests.swift
@@ -307,11 +307,11 @@ final class ChatClient_Tests: XCTestCase {
         )
         let connectionRepository = try XCTUnwrap(client.connectionRepository as? ConnectionRepository_Mock)
         connectionRepository.disconnectResult = .success(())
-        client.startTrackingChannelController(ChannelControllerSpy())
-        client.startTrackingChannelListController(ChatChannelListController_Mock.mock())
+        client.syncRepository.startTrackingChannelController(ChannelControllerSpy())
+        client.syncRepository.startTrackingChannelListController(ChatChannelListController_Mock.mock())
 
-        XCTAssertEqual(client.activeChannelControllers.count, 1)
-        XCTAssertEqual(client.activeChannelListControllers.count, 1)
+        XCTAssertEqual(client.syncRepository.activeChannelControllers.count, 1)
+        XCTAssertEqual(client.syncRepository.activeChannelListControllers.count, 1)
 
         // WHEN
         let expectation = self.expectation(description: "logout completes")
@@ -321,8 +321,8 @@ final class ChatClient_Tests: XCTestCase {
         waitForExpectations(timeout: defaultTimeout)
 
         // THEN
-        XCTAssertEqual(client.activeChannelControllers.count, 0)
-        XCTAssertEqual(client.activeChannelListControllers.count, 0)
+        XCTAssertEqual(client.syncRepository.activeChannelControllers.count, 0)
+        XCTAssertEqual(client.syncRepository.activeChannelListControllers.count, 0)
     }
 
     func test_apiClient_usesInjectedURLSessionConfiguration() {
@@ -824,69 +824,42 @@ final class ChatClient_Tests: XCTestCase {
 
         XCTAssertEqual(streamHeader, SystemEnvironment.xStreamClientHeader)
     }
-
-    // MARK: - Active Controller
-
-    func test_startTrackingChannelController() {
-        let client = ChatClient(config: .init())
-
-        let controller = ChatChannelController_Mock.mock()
-        client.startTrackingChannelController(controller)
-
-        XCTAssertTrue(client.activeChannelControllers.allObjects.first === controller)
+    
+    // MARK: - User Session
+    
+    func test_sharedCurrentUserController() {
+        let client = ChatClient_Mock.mock()
+        let controller1 = client.sharedCurrentUserController
+        let controller2 = client.sharedCurrentUserController
+        XCTAssertTrue(controller1 === controller2, "Shared instance should be returned")
     }
-
-    func test_startTrackingChannelController_whenAlreadyExists_thenDoNotDuplicate() {
-        let client = ChatClient(config: .init())
-
-        let controller = ChatChannelController_Mock.mock()
-        client.startTrackingChannelController(controller)
-        client.startTrackingChannelController(controller)
-
-        XCTAssertTrue(client.activeChannelControllers.allObjects.first === controller)
-        XCTAssertEqual(client.activeChannelControllers.allObjects.count, 1)
-    }
-
-    func test_stopTrackingChannelController() {
-        let client = ChatClient(config: .init())
-        let controller = ChatChannelController_Mock.mock()
-        client.startTrackingChannelController(controller)
-        XCTAssertEqual(client.activeChannelControllers.allObjects.count, 1)
-
-        client.stopTrackingChannelController(controller)
-
-        XCTAssertTrue(client.activeChannelControllers.allObjects.isEmpty)
-    }
-
-    func test_startTrackingChannelListController() {
-        let client = ChatClient(config: .init())
-
-        let controller = ChatChannelListController_Mock.mock()
-        client.startTrackingChannelListController(controller)
-
-        XCTAssertTrue(client.activeChannelListControllers.allObjects.first === controller)
-    }
-
-    func test_startTrackingChannelListController_whenAlreadyExists_thenDoNotDuplicate() {
-        let client = ChatClient(config: .init())
-
-        let controller = ChatChannelListController_Mock.mock()
-        client.startTrackingChannelListController(controller)
-        client.startTrackingChannelListController(controller)
-
-        XCTAssertTrue(client.activeChannelListControllers.allObjects.first === controller)
-        XCTAssertEqual(client.activeChannelListControllers.allObjects.count, 1)
-    }
-
-    func test_stopTrackingChannelListController() {
-        let client = ChatClient(config: .init())
-        let controller = ChatChannelListController_Mock.mock()
-        client.startTrackingChannelListController(controller)
-        XCTAssertEqual(client.activeChannelListControllers.allObjects.count, 1)
-
-        client.stopTrackingChannelListController(controller)
-
-        XCTAssertTrue(client.activeChannelListControllers.allObjects.isEmpty)
+    
+    func test_sharedCurrentUserController_whenConnectAndLogout_thenNewInstance() throws {
+        let client = ChatClient_Mock(config: inMemoryStorageConfig, environment: testEnv.environment)
+        let userInfo = UserInfo(id: "id1")
+        // Connect
+        let expectation = XCTestExpectation(description: "Connect")
+        client.mockAuthenticationRepository.connectUserResult = .success(())
+        try client.mockDatabaseContainer.createCurrentUser(id: userInfo.id)
+        client.connectUser(userInfo: userInfo, token: .unique()) { error in
+            guard error == nil else { return }
+            expectation.fulfill()
+        }
+        wait(for: [expectation], timeout: defaultTimeout)
+        let controller1 = client.sharedCurrentUserController
+        let controller2 = client.sharedCurrentUserController
+        XCTAssertTrue(controller1 === controller2, "Shared instance should be returned")
+        
+        // Logout
+        let expectation2 = XCTestExpectation(description: "Logout")
+        let connectionRepositoryMock = try XCTUnwrap(client.connectionRepository as? ConnectionRepository_Mock)
+        connectionRepositoryMock.disconnectResult = .success(())
+        client.logout {
+            expectation2.fulfill()
+        }
+        wait(for: [expectation2], timeout: defaultTimeout)
+        let controller3 = client.sharedCurrentUserController
+        XCTAssertFalse(controller1 === controller3, "New instance should be returned")
     }
 }