+++ Tests/StreamChatTests/Controllers/ChannelController/ChannelController_Tests.swift
@@ -36,6 +36,7 @@ final class ChannelController_Tests: XCTestCase {
     }
 
     override func tearDown() {
+        client?.cleanUp()
         env?.channelUpdater?.cleanUp()
         env?.eventSender?.cleanUp()
         env = nil
@@ -4597,12 +4598,12 @@ final class ChannelController_Tests: XCTestCase {
             channelListQuery: channelListQuery,
             client: client
         )
-        XCTAssert(client.activeChannelControllers.allObjects.isEmpty)
+        XCTAssert(client.syncRepository.activeChannelControllers.allObjects.isEmpty)
 
         controller.startWatching(isInRecoveryMode: false)
         XCTAssert(controller.client === client)
-        XCTAssert(client.activeChannelControllers.count == 1)
-        XCTAssert(client.activeChannelControllers.allObjects.first === controller)
+        XCTAssert(client.syncRepository.activeChannelControllers.count == 1)
+        XCTAssert(client.syncRepository.activeChannelControllers.allObjects.first === controller)
     }
 
     func test_startWatching_propagatesErrorFromUpdater() {
@@ -4783,10 +4784,10 @@ final class ChannelController_Tests: XCTestCase {
             client: client
         )
         controller.synchronize()
-        XCTAssert(client.activeChannelControllers.count == 1)
+        XCTAssert(client.syncRepository.activeChannelControllers.count == 1)
 
         controller.stopWatching()
-        XCTAssert(client.activeChannelControllers.allObjects.isEmpty == true)
+        XCTAssert(client.syncRepository.activeChannelControllers.allObjects.isEmpty == true)
     }
 
     // MARK: - Freeze channel
@@ -5137,8 +5138,8 @@ final class ChannelController_Tests: XCTestCase {
         controller.synchronize()
 
         XCTAssert(controller.client === client)
-        XCTAssert(client.activeChannelControllers.count == 1)
-        XCTAssert(client.activeChannelControllers.allObjects.first === controller)
+        XCTAssert(client.syncRepository.activeChannelControllers.count == 1)
+        XCTAssert(client.syncRepository.activeChannelControllers.allObjects.first === controller)
     }
 
     // MARK: shouldSendTypingEvents